+++ inject/inject-app/src/test/scala/com/twitter/inject/app/tests/FlagsModuleTest.scala
@@ -1,7 +1,7 @@
 package com.twitter.inject.app.tests
 
 import com.google.inject.{ConfigurationException, Guice, ProvisionException}
-import com.twitter.app.{Flaggable, Flags}
+import com.twitter.app.Flags
 import com.twitter.conversions.StorageUnitOps._
 import com.twitter.inject.annotations.{Flag, Flags => AnnotationFlags}
 import com.twitter.inject.app.internal.{FlagsModule, TwitterTypeConvertersModule}
@@ -9,14 +9,10 @@ import com.twitter.inject.{Injector, Test}
 import com.twitter.util.{Duration, StorageUnit, Time, TimeFormat}
 import java.net.InetSocketAddress
 import java.time.LocalTime
+import java.util.Arrays
 import javax.inject.Inject
 
 class FlagsModuleTest extends Test {
-  // we should consider adding this Flaggable to c.t.app.Flaggable
-  private implicit val ofJavaLocalTime: Flaggable[LocalTime] = new Flaggable[LocalTime] {
-    override def parse(s: String): LocalTime = LocalTime.parse(s)
-  }
-
   private[this] val flag =
     new Flags(this.getClass.getName, includeGlobal = false, failFastUntilParsed = false)
 
@@ -36,13 +32,49 @@ class FlagsModuleTest extends Test {
   private[this] val defaultStorageUnit = 2.gigabytes
   private[this] val defaultInetSocketAddress: InetSocketAddress =
     InetSocketAddress.createUnresolved("localhost", 0)
+
   flag[LocalTime]("local.time", defaultLocalTime, "A java.time.LocalTime")
   flag[Time]("twitter.time", defaultTime, "A twitter util Time")
   flag[Duration]("twitter.duration", defaultDuration, "A twitter util Duration")
   flag[StorageUnit]("storage.unit", defaultStorageUnit, "Represents a storage size")
   flag[InetSocketAddress]("some.address", defaultInetSocketAddress, "An InetSocketAddress")
 
-  flag.parseArgs(Array("-my.flag.value=bar"))
+  flag[Seq[String]]("seq.string", Seq.empty, "A Seq of String")
+  flag[Seq[LocalTime]]("seq.local.time", Seq(defaultLocalTime), "A Seq java.time.LocalTime")
+  flag[Seq[Time]]("seq.twitter.time", Seq(defaultTime), "A Seq of twitter util Time")
+  flag[Seq[Duration]](
+    "seq.twitter.duration",
+    Seq(defaultDuration),
+    "A Seq of twitter util Duration")
+  flag[Seq[StorageUnit]]("seq.storage.unit", Seq(defaultStorageUnit), "A Seq of storage size")
+  flag[Seq[InetSocketAddress]](
+    "seq.some.address",
+    Seq(defaultInetSocketAddress),
+    "A Seq of InetSocketAddress")
+
+  flag[java.util.List[String]]("list.string", Arrays.asList[String](), "A List of String")
+  flag[java.util.List[LocalTime]](
+    "list.local.time",
+    Arrays.asList(defaultLocalTime),
+    "A List java.time.LocalTime")
+  flag[java.util.List[Time]](
+    "list.twitter.time",
+    Arrays.asList(defaultTime),
+    "A List of twitter util Time")
+  flag[java.util.List[Duration]](
+    "list.twitter.duration",
+    Arrays.asList(defaultDuration),
+    "A List of twitter util Duration")
+  flag[java.util.List[StorageUnit]](
+    "list.storage.unit",
+    Arrays.asList(defaultStorageUnit),
+    "A List of storage size")
+  flag[java.util.List[InetSocketAddress]](
+    "list.some.address",
+    Arrays.asList(defaultInetSocketAddress),
+    "A List of InetSocketAddress")
+
+  flag.parseArgs(Array("-my.flag.value=bar", "-seq.string=foo,bar", "-list.string=baz,qux"))
 
   private[this] val flagsModule = new FlagsModule(flag)
   // Users should prefer the `c.t.inject.app.TestInjector`. Here we are testing the FlagsModule
@@ -104,6 +136,42 @@ class FlagsModuleTest extends Test {
       defaultStorageUnit)
     val inetSocketAddressFromInjector =
       injector.instance[InetSocketAddress](AnnotationFlags.named("some.address"))
+    inetSocketAddressFromInjector.getHostName should equal(defaultInetSocketAddress.getHostName)
+    inetSocketAddressFromInjector.getPort should equal(defaultInetSocketAddress.getPort)
+  }
+
+  test("test comma-separated type converters (Seq)") {
+    injector.instance[Seq[String]](AnnotationFlags.named("seq.string")) should equal(
+      Seq("foo", "bar"))
+    injector.instance[Seq[LocalTime]](AnnotationFlags.named("seq.local.time")) should equal(
+      Seq(defaultLocalTime))
+    injector.instance[Seq[Time]](AnnotationFlags.named("seq.twitter.time")) should equal(
+      Seq(defaultTime))
+    injector.instance[Seq[Duration]](AnnotationFlags.named("seq.twitter.duration")) should equal(
+      Seq(defaultDuration))
+    injector.instance[Seq[StorageUnit]](AnnotationFlags.named("seq.storage.unit")) should equal(
+      Seq(defaultStorageUnit))
+    val inetSocketAddressFromInjector =
+      injector.instance[Seq[InetSocketAddress]](AnnotationFlags.named("seq.some.address")).head
+    inetSocketAddressFromInjector.getHostName should equal(defaultInetSocketAddress.getHostName)
+    inetSocketAddressFromInjector.getPort should equal(defaultInetSocketAddress.getPort)
+  }
+
+  test("test comma-separated type converters (java.util.List)") {
+    injector.instance[java.util.List[String]](AnnotationFlags.named("list.string")) should equal(
+      Arrays.asList("baz", "qux"))
+    injector.instance[java.util.List[LocalTime]](
+      AnnotationFlags.named("list.local.time")) should equal(Arrays.asList(defaultLocalTime))
+    injector.instance[java.util.List[Time]](
+      AnnotationFlags.named("list.twitter.time")) should equal(Arrays.asList(defaultTime))
+    injector.instance[java.util.List[Duration]](
+      AnnotationFlags.named("list.twitter.duration")) should equal(Arrays.asList(defaultDuration))
+    injector.instance[java.util.List[StorageUnit]](
+      AnnotationFlags.named("list.storage.unit")) should equal(Arrays.asList(defaultStorageUnit))
+    val inetSocketAddressFromInjector =
+      injector
+        .instance[java.util.List[InetSocketAddress]](
+          AnnotationFlags.named("list.some.address")).get(0)
     inetSocketAddressFromInjector.getHostName should equal(defaultInetSocketAddress.getHostName)
     inetSocketAddressFromInjector.getPort should equal(defaultInetSocketAddress.getPort)
   }